:white_check_mark: Update system tests fixtures

+++ system-tests-fixtures/ios-container/ElectrodeContainer/Libraries/ReactNative/React/Modules/RCTEventEmitter.h
@@ -45,6 +45,6 @@
 - (void)stopObserving;
 
 - (void)addListener:(NSString *)eventName;
-- (void)removeListeners:(NSInteger)count;
+- (void)removeListeners:(double)count;
 
 @end
